backend/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.119.0
fastapi-login==1.10.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic[email]==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1